Move AutomationList up one level
